+++ app/build.gradle
@@ -2,6 +2,14 @@ apply plugin: 'com.android.application'
 apply plugin: 'com.google.gms.google-services'
 
 android {
+    signingConfigs {
+        config {
+            keyAlias 'key_for_debug'
+            keyPassword 'unsafex2'
+            storePassword 'unsafe'
+            storeFile file('../keystore.jks')
+        }
+    }
     compileSdkVersion 28
     defaultConfig {
         applicationId "com.taller2.droidclient"
@@ -14,23 +22,24 @@ android {
     buildTypes {
         debug {
             debuggable true
+            signingConfig signingConfigs.config
         }
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+            signingConfig signingConfigs.config
         }
     }
 }
 
 dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation fileTree(include: ['*.jar'], dir: 'libs')
     implementation 'com.android.support:appcompat-v7:28.0.0'
     implementation 'com.android.support.constraint:constraint-layout:1.1.3'
     implementation 'com.android.volley:volley:1.1.1'
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
-
     //For firebase
     implementation 'com.android.support:support-v13:28.0.0'
     implementation 'com.android.support:design:28.0.0'
@@ -40,10 +49,8 @@ dependencies {
     implementation 'com.google.firebase:firebase-auth:10.2.6'
     implementation 'com.google.firebase:firebase-database:10.2.6'
     implementation 'com.google.firebase:firebase-core:10.2.6'
-
+    implementation 'com.facebook.android:facebook-android-sdk:[4,5)'
     implementation 'com.android.support:cardview-v7:28.0.0'
-
     implementation 'com.google.android.gms:play-services-auth:10.2.6'
-
     implementation 'com.rengwuxian.materialedittext:library:2.1.4'
 }
