+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,5 @@
 <resources>
     <string name="app_name">DroidClient</string>
+    <string name="facebook_app_id">613082942449635</string>
+    <string name="fb_login_protocol_scheme">fb613082942449635</string>
 </resources>
